add testing-library support, updated dependencies, minor fixes and code cleanup
