+++ package.json
@@ -6,8 +6,8 @@
     "devDependencies": {
         "bluebird": "3.7.2",
         "common-readme": "1.1.0",
-        "fs-extra": "8.1.0",
-        "tar": "6.0.1"
+        "fs-extra": "9.0.1",
+        "tar": "6.0.5"
     },
     "scripts": {
         "build": "bash scripts/build-plugin.bash",
