+++ react-app/package.json
@@ -3,44 +3,56 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@ant-design/icons": "4.2.2",
-        "@kbase/ui-components": "0.2.1-alpha.21",
+        "@ant-design/icons": "4.3.0",
+        "@kbase/ui-components": "0.2.1-alpha.22",
         "@kbase/ui-lib": "0.2.1-alpha.3",
+        "@testing-library/react": "11.2.2",
+        "@testing-library/jest-dom": "5.11.8",
         "@types/jest": "*",
-        "@types/marked": "*",
-        "@types/node": "*",
-        "@types/react": "*",
-        "@types/react-dom": "*",
-        "@types/react-redux": "*",
-        "antd": "4.6.6",
-        "antd-dayjs-webpack-plugin": "1.0.1",
-        "axios": "0.20.0",
-        "babel-plugin-import": "1.13.0",
+        "@types/marked": "1",
+        "@types/node": "12",
+        "@types/react": "16",
+        "@types/react-dom": "16",
+        "@types/react-redux": "7",
+        "@types/testing-library__react": "10.2.0",
+        "@types/testing-library__jest-dom": "5.9.5",
+        "antd": "4.10.0",
+        "antd-dayjs-webpack-plugin": "1.0.4",
+        "axios": "0.21.1",
+        "babel-plugin-import": "1.13.3",
         "customize-cra": "1.0.0",
-        "less": "3.12.2",
-        "less-loader": "7.0.1",
-        "marked": "1.2.0",
+        "less": "4.0.0",
+        "less-loader": "7.2.1",
+        "marked": "1.2.7",
         "morgan": "1.10.0",
-        "react": "16.13.1",
-        "react-app-rewired": "2.1.6",
-        "react-dom": "16.13.1",
-        "react-redux": "7.2.1",
-        "react-scripts": "3.4.3",
+        "react": "16.14.0",
+        "react-app-rewired": "2.1.8",
+        "react-dom": "16.14.0",
+        "react-redux": "7.2.2",
+        "react-scripts": "4.0.1",
         "redux": "4.0.5",
         "redux-thunk": "2.3.0",
-        "vis-data": "7.0.0",
-        "vis-network": "8.3.2"
+        "vis-data": "7.1.1",
+        "vis-network": "8.5.6"
     },
     "devDependencies": {
-        "http-proxy-middleware": "1.0.5",
-        "nodemon": "2.0.4",
-        "typescript": "3.9.7"
+        "http-proxy-middleware": "1.0.6",
+        "nodemon": "2.0.6",
+        "typescript": "4.1.3"
     },
     "scripts": {
         "start": "nodemon -w ./config-overrides.js --exec 'react-app-rewired start'",
         "startx": "react-app-rewired start",
         "build": "react-app-rewired build",
-        "test": "react-app-rewired test"
+        "test": "react-app-rewired test",
+        "test-coverage": "react-app-rewired test --coverage --watchAll=false"
+    },
+    "jest": {
+        "collectCoverageFrom": [
+            "src/**/*.ts",
+            "src/**/*.tsx",
+            "!**/*.js"
+        ]
     },
     "eslintConfig": {
         "extends": "react-app"
