+++ react-app/src/types/index.ts
@@ -27,3 +27,10 @@ export interface ErrorState {
 export interface LoadedState {
     status: LoadingStatus.LOADED;
 }
+
+
+export interface SimpleError {
+    message: string;
+}
+
+
