+++ react-app/src/ui/ErrorView.tsx
@@ -4,6 +4,7 @@ import { Alert } from 'antd';
 import { ExclamationCircleOutlined } from "@ant-design/icons";
 
 export interface ErrorViewProps {
+    title?: string;
     error: UIError;
 }
 
@@ -14,12 +15,15 @@ interface ErrorViewState {
 export default class ErrorView extends React.Component<ErrorViewProps, ErrorViewState> {
     renderMessage() {
         return <>
-            <p>Error!</p>
+            <p>{this.props.title || 'Error!'}</p>
             <p>
                 <ExclamationCircleOutlined style={{ color: 'red' }} />
                 {' '}
                 {this.props.error.message}
             </p>
+
+            <p>Code: <span>{this.props.error.code}</span></p>
+            <p>Source: <span>{this.props.error.source}</span></p>
         </>;
     }
     render() {
